Integrate Travis CI (#1)
* Initial push for Travis.
* Investigated if Travis lets us do privileged containers (KVM) - it doesn't seem that way.
* Cleaned up script naming.
* Added Perl dependencies to cpanfile.
* Switch to /usr/bin/env perl instead of the system global one. (Compat with travis cpanm)
* Make test-all.sh take an argument to match which OS:es should be tested.
* Parallelized builds using Travis
* Store interim TAP results to S3 bucket
* Fix a typo in the regexp for ensuring MD-style tables.
* Generate (from S3 bucket data) and publish report to Github Wiki.
* Also publish README.md as Home.md -> Wiki repo.

+++ test/test-all.sh
@@ -1,48 +1,54 @@
 #!/bin/bash
 
+REQUESTED="${1:-}"
+
 # Test "regular" docker image-based OS distributions:
 for VER in madworx/debian-archive:{etch,lenny,squeeze} \
            debian:{buster,jessie,sid,stretch,wheezy}-slim \
            ubuntu:{artful,bionic,trusty,xenial,zesty,yakkety,precise} \
            centos:{6,7} \
            opensuse/{leap,tumbleweed} ; do
-    echo "Testing image ${VER}" 1>&2
-    docker run --rm \
-           -v $(dirname $(readlink -f $0))/..:/docshell \
-           -i \
-           -e VER \
-           "${VER}" \
-           /docshell/test/test-shells.sh "${VER}"
+    if [[ "${VER}" = *${REQUESTED}* ]] ; then
+        echo "Testing image ${VER}" 1>&2
+        docker run --rm \
+               -v $(dirname $(readlink -f $0))/..:/docshell \
+               -i \
+               -e VER \
+               "${VER}" \
+               /docshell/test/test-shells.sh "${VER}"
+    fi
 done
-
 
 # Test Alpine Linux, who has special needs:
 for VER in alpine:3.{1,2,3,4,5,6,7} alpine:edge ; do
-    echo "Testing image ${VER}" 1>&2
-    docker run --rm \
-           -v $(dirname $(readlink -f $0))/..:/docshell \
-           -i \
-           -e VER "${VER}" \
-           /bin/sh -c "apk update >/dev/null 2>&1 ; \
-                      apk add bash >/dev/null 2>&1 ; \
-                      /docshell/test/test-shells.sh \"${VER}\""
+    if [[ "${VER}" = *${REQUESTED}* ]] ; then
+        echo "Testing image ${VER}" 1>&2
+        docker run --rm \
+               -v $(dirname $(readlink -f $0))/..:/docshell \
+               -i \
+               -e VER "${VER}" \
+               /bin/sh -c "apk update >/dev/null 2>&1 ; \
+                           apk add bash >/dev/null 2>&1 ; \
+                           /docshell/test/test-shells.sh \"${VER}\""
+    fi
 done
 
-
-# Test NetBSD, we need a prefix to the volume mount
+# Test NetBSD, we need a prefix to the volume mount:
 for VER in madworx/netbsd:{7.1.2,6.1.5}-x86_64 ; do
-    echo "Testing image ${VER}" 1>&2
-    docker run --rm \
-           -v $(dirname $(readlink -f $0))/..:/bsd/docshell \
-           -i \
-           -e VER \
-           "${VER}" \
-           /docshell/test/test-shells.sh "${VER}"
+    if [[ "${VER}" = *${REQUESTED}* ]] ; then
+        echo "Testing image ${VER}" 1>&2
+        docker run --privileged --rm \
+               -v $(dirname $(readlink -f $0))/..:/bsd/docshell \
+               -i \
+               -e VER \
+               "${VER}" \
+               /docshell/test/test-shells.sh "${VER}"
+    fi
 done
 
-
-# Test "multibash"
-docker run --rm -i madworx/multibash /bin/bash -c ' \
+# Test "multibash":
+if [[ "bash" = *${REQUESTED}* ]] ; then
+    docker run --rm -i madworx/multibash /bin/bash -c ' \
 base64 -d | tar xf - ; \
 os="bash"
 for SHL in /usr/local/bin/bash-* ; do
@@ -55,3 +61,4 @@ for SHL in /usr/local/bin/bash-* ; do
     rm /tmp/tmp.output
     echo ""
 done' < <(cd ../ ; tar cf - example.sh test/example.expect | base64)
+fi
